+++ Package.swift
@@ -10,23 +10,29 @@ let package = Package(
     ],
     products: [
         .library(
-            name: "YandexMobileAdsPackage",
-            targets: ["YandexMobileAdsTarget"]
+            name: "YandexMobileAds",
+            targets: ["YandexMobileAdsWrapper"]
+        ),
+        .library(
+            name: "YandexMobileAdsInstream",
+            targets: ["YandexMobileAdsInstreamWrapper"]
         )
     ],
     dependencies: [
         .package(url: "https://github.com/appmetrica/appmetrica-sdk-ios", "5.0.0"..<"6.0.0"),
-        .package(url: "https://github.com/divkit/divkit-ios", exact: "30.7.0")
+        .package(url: "https://github.com/divkit/divkit-ios", exact: "30.11.0")
     ],
     targets: [
         .target(
-            name: "YandexMobileAdsTarget",
+            name: "YandexMobileAdsWrapper",
             dependencies: [
                 .target(name: "YandexMobileAds"),
-                .target(name: "YandexMobileAdsInstream"),
                 .product(name: "AppMetricaCore", package: "appmetrica-sdk-ios"),
                 .product(name: "AppMetricaCrashes", package: "appmetrica-sdk-ios"),
                 .product(name: "DivKit", package: "divkit-ios")
+            ],
+            resources: [
+                .copy("Resources/MobileAdsBundle.bundle")
             ],
             linkerSettings: [
                 .linkedFramework("AVFoundation"),
@@ -48,15 +54,22 @@ let package = Package(
                 .linkedFramework("WebKit")
             ]
         ),
+        .target(
+            name: "YandexMobileAdsInstreamWrapper",
+            dependencies: [
+                .target(name: "YandexMobileAdsInstream"),
+                .target(name: "YandexMobileAdsWrapper")
+            ]
+        ),
         .binaryTarget(
             name: "YandexMobileAds",
-            url: "https://ads-mobile-sdk.s3.yandex.net/Yandex/YandexMobileAds/7.3.0/spm/95bb0894-5077-4715-a965-de89f33adb87.zip",
-            checksum: "590427af9146dc73b74c6fcaa2d19992cbeeb856af3b161936910e9ade6c2b61"
+            url: "https://ads-mobile-sdk.s3.yandex.net/Yandex/YandexMobileAds/7.3.1/spm/ff069641-e2f8-4667-9548-66799a131ba0.zip",
+            checksum: "709bf7d8dfa561b81bc07476f44779628b32160b6daec863b046635ee5d232a7"
         ),
         .binaryTarget(
             name: "YandexMobileAdsInstream",
-            url: "https://ads-mobile-sdk.s3.yandex.net/YandexMobileAdsInstream/0.32.0/spm/b81f9f29-138e-4827-ae69-48fd8709a76a.zip",
-            checksum: "d271d5bd881be6ca9530a04d37d0a86d31a68743d4d35bcec5a15826fdd2a1fe"
+            url: "https://ads-mobile-sdk.s3.yandex.net/YandexMobileAdsInstream/0.33.0/spm/1f7ed375-39f4-4506-b0f4-6f51c3f0e7be.zip",
+            checksum: "6cd5c40908979ef00ea5bc8bcd498e554c81bfb985df71e8a3e610ff12bb1323"
         )
     ]
 )