+++ Sources/YandexMobileAdsWrapper/MACSwiftBundleProvider.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+@objc(MACSwiftBundleProvider)
+class MACSwiftBundleProvider: NSObject {
+
+    @objc func provideBundle() -> Bundle {
+        Bundle.module
+    }
+}