+++ changelog/mediation/mobileads-chartboost/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version 9.7.0.7
+
+### Updated
+
+- Updated minimum supported `YandexMobileAds` version to `7.3.1`
+
 ## Version 9.7.0.6
 
 ### Updated