+++ changelog/mediation/mobileads-startapp/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version 4.10.3.4
+
+### Updated
+
+- Updated minimum supported `YandexMobileAds` version to `7.3.1`
+
 ## Version 4.10.3.3
 
 ### Updated